requirements.txt
speechmatics-python==1.11.1
gradio==3.41.2
openai==1.3.4
